Crate common_greptimedb_telemetry

Structs

GreptimeDBTelemetry

Report version info to GreptimeDB. We do not collect any identity-sensitive information. This task is scheduled to run every 30 minutes. The task will be disabled default. It can be enabled by setting the build feature greptimedb-telemetry Collector is used to collect the version info. It can be implemented by different components. client is used to send the HTTP request to GreptimeDB. telemetry_url is the GreptimeDB url.

StatisticData 🔒

Telemetry data to report

Enums

GreptimeDBTelemetryTask

Mode

Constants

GREPTIMEDB_TELEMETRY_CLIENT_CONNECT_TIMEOUT 🔒

The default connect timeout to greptime cloud.

GREPTIMEDB_TELEMETRY_CLIENT_REQUEST_TIMEOUT 🔒

The default request timeout to greptime cloud.

TELEMETRY_URL

The URL to report telemetry data.

UUID_FILE_NAME 🔒

The local installation uuid cache file

Statics

TELEMETRY_INTERVAL

The default interval of reporting telemetry data to greptime cloud

Traits

Collector

Functions

default_get_uuid

print_anonymous_usage_data_disclaimer 🔒