common_greptimedb_telemetry

Struct GreptimeDBTelemetry

pub struct GreptimeDBTelemetry {
    pub(crate) statistics: Box<dyn Collector + Send + Sync>,
    pub(crate) client: Option<Client>,
    pub(crate) working_home: Option<String>,
    pub(crate) telemetry_url: &'static str,
    pub(crate) should_report: Arc<AtomicBool>,
    pub(crate) report_times: usize,
}

Report version info to GreptimeDB.

We do not collect any identity-sensitive information. This task is scheduled to run every 30 minutes. The task will be disabled default. It can be enabled by setting the build feature greptimedb-telemetry Collector is used to collect the version info. It can be implemented by different components. client is used to send the HTTP request to GreptimeDB. telemetry_url is the GreptimeDB url.

Fields
statistics: Box<dyn Collector + Send + Sync>client: Option<Client>working_home: Option<String>telemetry_url: &'static strshould_report: Arc<AtomicBool>report_times: usize
Implementations

impl GreptimeDBTelemetry

pub fn new( working_home: Option<String>, statistics: Box<dyn Collector + Send + Sync>, should_report: Arc<AtomicBool>, ) -> Self

pub async fn report_telemetry_info(&mut self) -> Option<Response>